requirements.txt
fastapi==0.70.0
requests
uvicorn==0.15.0
aiohttp==3.7.4.post0
aioredis==2.0.0
motor==2.5.1
pydantic==1.8.2
pymongo==3.12.1
pytz==2021.3
semantic-version==2.8.5
beanie==1.7.0
pytest==6.2.5
fastapi-cache2==0.1.7